# swagger.yaml
openapi: 3.0.4
info:
  title: LAPEN Agenda API
  description: Tennis Court Management System with Betting Integration
  version: 1.0.0
  contact:
    email: lapen.ptc@gmail.com
servers:
  - url: http://localhost:5001
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminSession:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    Tournament:
      type: object
      properties:
        id: {type: integer, example: 1}
        name: {type: string, example: "Copa LAPEN 2026"}
        slug: {type: string, example: "copa-lapen-2026"}
        description: {type: string, nullable: true}
        location: {type: string, nullable: true}
        start_date: {type: string, format: date, example: "2026-11-10"}
        end_date: {type: string, format: date, example: "2026-11-12"}
        registration_opens_at: {type: string, format: date-time, nullable: true, description: "Brazil local time, no timezone"}
        registration_closes_at: {type: string, format: date-time, nullable: true, description: "Brazil local time, no timezone"}
        rules_text: {type: string, nullable: true}
        contact_info: {type: string, nullable: true}
        match_format:
          type: string
          enum: [best_of_3_super_tb, pro_set_8, single_set_6]
          description: Chosen when the tournament is created, locked after the first result
        no_ad: {type: boolean}
        match_tiebreak_points: {type: integer, example: 10}
        status:
          type: string
          enum: [draft, registration_open, registration_closed, in_progress, finished, cancelled]
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    TournamentCategory:
      type: object
      properties:
        id: {type: integer}
        tournament_id: {type: integer}
        name: {type: string, example: "Masculino 3a Classe"}
        max_entries: {type: integer, nullable: true, description: null means no limit}
        min_entries: {type: integer, example: 4}
        draw_format: {type: string, enum: [knockout, round_robin, groups_knockout]}
        group_target_size: {type: integer, nullable: true, enum: [3, 4]}
        qualifiers_per_group: {type: integer, nullable: true, enum: [1, 2]}
        num_seeds: {type: integer}
        wo_tolerance_min: {type: integer, example: 15}
        min_rest_min: {type: integer, example: 30}
        status:
          type: string
          enum: [awaiting_draw, drawn, published, group_stage, knockout_stage, finished]
        sort_order: {type: integer}
        counts:
          type: object
          description: Registrations per status (only on the tournament detail)
          additionalProperties: {type: integer}
    TournamentRegistration:
      type: object
      description: Admin view. Contact data is never exposed by public endpoints.
      properties:
        id: {type: integer}
        category_id: {type: integer}
        user_id: {type: integer, nullable: true, description: Set only for approved LAPEN members}
        is_member: {type: boolean}
        full_name: {type: string}
        display_name: {type: string}
        email: {type: string}
        phone: {type: string, description: Digits only}
        notes: {type: string, nullable: true}
        status:
          type: string
          enum: [pending, confirmed, rejected, cancelled, waitlist, withdrawn]
        seed: {type: integer, nullable: true}
        rejection_reason: {type: string, nullable: true}
        terms_accepted_at: {type: string, format: date-time}
        data_consent_at: {type: string, format: date-time}
        reviewed_at: {type: string, format: date-time, nullable: true}
        created_at: {type: string, format: date-time}
    TournamentError:
      type: object
      properties:
        error: {type: string}
        code: {type: string, description: "Stable machine-readable reason", example: active_tournament_exists}
...
    PublicTournament:
      type: object
      description: Public view. Drafts and cancelled tournaments are not public.
      properties:
        id: {type: integer}
        name: {type: string}
        slug: {type: string}
        description: {type: string, nullable: true}
        location: {type: string, nullable: true}
        start_date: {type: string, format: date}
        end_date: {type: string, format: date}
        rules_text: {type: string, nullable: true}
        contact_info: {type: string, nullable: true}
        match_format: {type: string, enum: [best_of_3_super_tb, pro_set_8, single_set_6]}
        no_ad: {type: boolean}
        match_tiebreak_points: {type: integer}
        format_text: {type: string, example: "2 sets sem vantagem + super tie-break de 10 pts"}
        status: {type: string, enum: [registration_open, registration_closed, in_progress, finished]}
        registration:
          type: object
          properties:
            open: {type: boolean, description: True while the sign-up page accepts registrations}
            state: {type: string, enum: [not_started, open, closed]}
            opens_at: {type: string, format: date-time, nullable: true}
            closes_at: {type: string, format: date-time, nullable: true}
    PublicEntry:
      type: object
      nullable: true
      description: A player. Only the display name and seed are ever public.
      properties:
        id: {type: integer}
        display_name: {type: string}
        seed: {type: integer, nullable: true}
        withdrawn: {type: boolean}
    PublicMatch:
      type: object
      properties:
        id: {type: integer}
        stage: {type: string, enum: [group, knockout]}
        group: {type: string, nullable: true, example: A}
        round_number: {type: integer}
        round_name: {type: string, description: Knockout only}
        position: {type: integer}
        sides:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: object
            properties:
              entry: {$ref: '#/components/schemas/PublicEntry'}
              source: {type: string, nullable: true, description: "Where an open slot comes from, e.g. 1o Grupo A, Vencedor das quartas 2 or Sem adversario (duplo W.O.). An empty side without source on a bye match is the BYE"}
        status: {type: string, enum: [pending, completed]}
        outcome: {type: string, nullable: true, enum: [normal, wo, double_wo, retired, bye]}
        score: {type: string, nullable: true}
        winner_side: {type: integer, nullable: true, enum: [1, 2]}
        played_at: {type: string, format: date-time, nullable: true}
        planned_date: {type: string, format: date, nullable: true}
        planned_time: {type: string, nullable: true, example: "09:30"}
        court: {type: string, nullable: true}
    PublicCategorySummary:
      type: object
      properties:
        id: {type: integer}
        name: {type: string}
        draw_format: {type: string, enum: [knockout, round_robin, groups_knockout]}
        status: {type: string, enum: [awaiting_draw, published, group_stage, knockout_stage, finished], description: A draw that is only a preview shows as awaiting_draw}
        stage:
          type: object
          properties:
            current: {type: string, enum: [registration, groups, knockout, final, champion]}
            steps:
              type: array
              items:
                type: object
                properties:
                  key: {type: string}
                  label: {type: string, example: "Mata-mata"}
                  state: {type: string, enum: [done, current, upcoming]}
        progress:
          type: object
          description: Matches with a result over matches to be played. Byes are not matches. Phases not built yet are not counted
          properties:
            done: {type: integer}
            total: {type: integer}
            group: {type: object, properties: {done: {type: integer}, total: {type: integer}}}
            knockout: {type: object, properties: {done: {type: integer}, total: {type: integer}}}
        capacity:
          type: object
          properties:
            confirmed: {type: integer}
            max_entries: {type: integer, nullable: true}
            min_entries: {type: integer}
        group_target_size: {type: integer, nullable: true}
        qualifiers_per_group: {type: integer, nullable: true}
        num_seeds: {type: integer}
        wo_tolerance_min: {type: integer}
        min_rest_min: {type: integer}
        draw_published: {type: boolean}
        champion: {$ref: '#/components/schemas/PublicEntry'}
        runner_up: {$ref: '#/components/schemas/PublicEntry'}
    PublicCategory:
      type: object
      description: Everything the tracking screen needs for one category, in a single response.
      properties:
        tournament: {$ref: '#/components/schemas/PublicTournament'}
        category: {$ref: '#/components/schemas/PublicCategorySummary'}
        entries:
          type: array
          description: Confirmed registrations only, seeds first
          items: {$ref: '#/components/schemas/PublicEntry'}
        groups:
          type: array
          items:
            type: object
            properties:
              id: {type: integer}
              name: {type: string}
              complete: {type: boolean}
              blocked: {type: boolean}
              confirmed: {type: boolean}
              qualifiers: {type: integer}
              ties:
                type: array
                items:
                  type: object
                  properties:
                    entries:
                      type: array
                      items: {$ref: '#/components/schemas/PublicEntry'}
                    relevant: {type: boolean}
              rows:
                type: array
                items:
                  type: object
                  properties:
                    entry: {$ref: '#/components/schemas/PublicEntry'}
                    position: {type: integer}
                    tied: {type: boolean}
                    state: {type: string, enum: [provisional, open, qualified, eliminated, tie_pending]}
                    played: {type: integer}
                    wins: {type: integer}
                    losses: {type: integer}
                    sets_won: {type: integer}
                    sets_lost: {type: integer}
                    sets_diff: {type: integer}
                    games_won: {type: integer}
                    games_lost: {type: integer}
                    games_diff: {type: integer}
                    destination: {type: string, nullable: true, example: "Semifinal 1", description: Where a qualified player goes next}
              matches:
                type: array
                items: {$ref: '#/components/schemas/PublicMatch'}
        bracket:
          type: object
          nullable: true
          properties:
            size: {type: integer}
            rounds:
              type: array
              items:
                type: object
                properties:
                  round_number: {type: integer}
                  name: {type: string}
                  matches:
                    type: array
                    items: {$ref: '#/components/schemas/PublicMatch'}
        knockout_pending: {type: boolean, description: "True when the knockout phase will be drawn at the end of the groups"}
        upcoming:
          type: array
          description: Up to 5 games ready to be played, scheduled ones first
          items: {$ref: '#/components/schemas/PublicMatch'}
        results:
          type: array
          description: Up to 5 most recent results
          items: {$ref: '#/components/schemas/PublicMatch'}
...
paths:
...
  /api/admin/tournaments:
    get:
      tags: [Tournaments]
      summary: List tournaments (active first) with category and pending-registration counts
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Tournaments
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournaments:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Tournament'
                        - type: object
                          properties:
                            categories_count: {type: integer}
                            pending_registrations: {type: integer}
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
    post:
      tags: [Tournaments]
      summary: Create a tournament (always starts as draft)
      description: The slug is generated from the name unless given. Defaults match_format best_of_3_super_tb, no_ad true, match_tiebreak_points 10.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, start_date, end_date]
              properties:
                name: {type: string}
                slug: {type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'}
                description: {type: string}
                location: {type: string}
                start_date: {type: string, format: date}
                end_date: {type: string, format: date}
                registration_opens_at: {type: string, example: "2026-11-01T08:00"}
                registration_closes_at: {type: string, example: "2026-11-08T18:00"}
                rules_text: {type: string}
                contact_info: {type: string}
                match_format: {type: string, enum: [best_of_3_super_tb, pro_set_8, single_set_6]}
                no_ad: {type: boolean}
                match_tiebreak_points: {type: integer}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament: {$ref: '#/components/schemas/Tournament'}
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Slug already taken (code slug_taken)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
  /api/admin/tournaments/{tournament_id}:
    parameters:
      - {name: tournament_id, in: path, required: true, schema: {type: integer}}
    get:
      tags: [Tournaments]
      summary: Tournament with its categories and registration counts
      description: Registrations whose deadline passed are closed lazily when this is read.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament: {$ref: '#/components/schemas/Tournament'}
                  categories:
                    type: array
                    items: {$ref: '#/components/schemas/TournamentCategory'}
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
    put:
      tags: [Tournaments]
      summary: Update a tournament (partial)
      description: The slug never changes. Finished or cancelled tournaments are read-only. match_format is locked after the first result (code format_locked).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Any subset of the fields accepted on creation, except slug
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament: {$ref: '#/components/schemas/Tournament'}
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Read-only or format locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
    delete:
      tags: [Tournaments]
      summary: Delete a draft tournament (cascades to categories)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Not a draft (code not_draft)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
...
  /api/admin/tournaments/{tournament_id}/categories/{category_id}:
    parameters:
      - {name: tournament_id, in: path, required: true, schema: {type: integer}}
      - {name: category_id, in: path, required: true, schema: {type: integer}}
    put:
      tags: [Tournaments]
      summary: Update a category (partial)
      description: Format, group settings and num_seeds are locked once the draw exists (code draw_exists). max_entries cannot drop below the confirmed count and num_seeds cannot drop below an assigned seed.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Any subset of the fields accepted on creation
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  category: {$ref: '#/components/schemas/TournamentCategory'}
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Locked, below_confirmed, seed_assigned or duplicate name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
    delete:
      tags: [Tournaments]
      summary: Delete a category without registrations and without a draw
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Has registrations or draw exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
  /api/admin/tournaments/{tournament_id}/registrations:
    parameters:
      - {name: tournament_id, in: path, required: true, schema: {type: integer}}
    get:
      tags: [Tournaments]
      summary: List registrations with per-category counts
      description: Includes e-mail and phone (admin only). Search is accent-insensitive on name, display name, e-mail and phone.
      security:
        - BearerAuth: []
      parameters:
        - {name: category_id, in: query, schema: {type: integer}}
        - {name: status, in: query, schema: {type: string, enum: [pending, confirmed, rejected, cancelled, waitlist, withdrawn]}}
        - {name: q, in: query, schema: {type: string}}
      responses:
        '200':
          description: Registrations
          content:
            application/json:
              schema:
                type: object
                properties:
                  registrations:
                    type: array
                    items: {$ref: '#/components/schemas/TournamentRegistration'}
                  counts:
                    type: object
                    description: category id to registrations per status
                    additionalProperties:
                      type: object
                      additionalProperties: {type: integer}
        '400':
          description: Invalid status filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
    post:
      tags: [Tournaments]
      summary: Register someone manually (for example received by WhatsApp)
      description: Starts as confirmed unless status is pending. Respects the category capacity. Optional user_id links an approved LAPEN member.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category_id, full_name, email, phone]
              properties:
                category_id: {type: integer}
                full_name: {type: string}
                display_name: {type: string}
                email: {type: string}
                phone: {type: string}
                notes: {type: string}
                status: {type: string, enum: [pending, confirmed], default: confirmed}
                user_id: {type: integer}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  registration: {$ref: '#/components/schemas/TournamentRegistration'}
        '400':
          description: Invalid data or user is not an approved LAPEN member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Category full, duplicate e-mail or member, draw exists, tournament locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
...
  /api/tournaments/{slug}/registrations:
    post:
      tags: [Tournaments]
      summary: Public tournament sign-up
      description: |
        No login needed. With a valid token of an approved LAPEN member the registration is linked to that member.
        The registration starts as pending, or waitlist when the category already has all its places confirmed.
        At most 5 sign-ups per hour per client address. The response never contains contact data.
        The field website is a honeypot: when filled the API answers 201 and stores nothing.
      parameters:
        - {name: slug, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category_id, full_name, email, phone, accepted_terms, data_consent]
              properties:
                category_id: {type: integer}
                full_name: {type: string}
                display_name: {type: string, description: Defaults to the first two names}
                email: {type: string}
                phone: {type: string, description: "With area code, 10 to 13 digits"}
                notes: {type: string, maxLength: 500}
                accepted_terms: {type: boolean, description: Must be true}
                data_consent: {type: boolean, description: Must be true}
                website: {type: string, description: "Honeypot, leave empty"}
      responses:
        '201':
          description: Received
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: {type: string}
                  waitlisted: {type: boolean}
                  registration:
                    type: object
                    nullable: true
                    properties:
                      id: {type: integer}
                      category_id: {type: integer}
                      display_name: {type: string}
                      status: {type: string, enum: [pending, waitlist]}
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '404':
          description: Tournament, category or draft not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '409':
          description: Registrations not open yet or closed (codes registration_not_started, registration_closed), duplicate e-mail or member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
        '429':
          description: Rate limited (code rate_limited)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
...
  /api/tournaments/{slug}:
    parameters:
      - {name: slug, in: path, required: true, schema: {type: string}}
    get:
      tags: [Tournaments]
      summary: A tournament with its categories (stage, progress and capacity of each)
      responses:
        '200':
          description: Tournament
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament: {$ref: '#/components/schemas/PublicTournament'}
                  categories:
                    type: array
                    items: {$ref: '#/components/schemas/PublicCategorySummary'}
                  progress: {type: object, properties: {done: {type: integer}, total: {type: integer}}}
        '404':
          description: Unknown, draft or cancelled tournament
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
  /api/tournaments/{slug}/categories/{category_id}:
    parameters:
      - {name: slug, in: path, required: true, schema: {type: string}}
      - {name: category_id, in: path, required: true, schema: {type: integer}}
    get:
      tags: [Tournaments]
      summary: One category for the tracking screen (groups, bracket, games and results in one request)
      description: Before the draw is published, groups is empty and bracket is null. Contact data never appears.
      responses:
        '200':
          description: Category
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PublicCategory'}
        '404':
          description: Unknown category, or not public
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
  /api/tournaments/{slug}/matches:
    parameters:
      - {name: slug, in: path, required: true, schema: {type: string}}
    get:
      tags: [Tournaments]
      summary: Upcoming games or results of the whole tournament
      description: Upcoming = both players known and no result, scheduled ones first by date and time (times only once the schedule is published). Results = most recent first. Up to 200.
      parameters:
        - {name: view, in: query, required: true, schema: {type: string, enum: [upcoming, results]}}
        - {name: category, in: query, schema: {type: integer}}
        - {name: stage, in: query, schema: {type: string, enum: [group, knockout]}}
      responses:
        '200':
          description: Matches
          content:
            application/json:
              schema:
                type: object
                properties:
                  view: {type: string}
                  matches:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/PublicMatch'
                        - type: object
                          properties:
                            category: {type: object, properties: {id: {type: integer}, name: {type: string}}}
        '400':
          description: Missing or invalid view or stage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentError'
...